+++ utils/rx-page-objects/index.d.ts
@@ -25,8 +25,7 @@ export {rxSelect, rxSelectAccessor} from './src/rxSelect.page';
 // export {rxSortableColumn, SORT_TYPE} from './src/rxSortableColumn.page';
 // export {rxStatusColumn, STATUS_COLORS, STATUS_ICONS} from './src/rxStatusColumn.page';
 // export {rxTags, Tag} from './src/rxTags.page';
-export {rxTimePicker} from './src/rxTimePicker.page';
-export {rxTimePickerUtil} from './src/rxTimePickerUtil.page';
+export {parseUtcOffset, rxTimePicker} from './src/rxTimePicker.page';
 // export {rxToggleSwitch} from './src/rxToggleSwitch.page';
 // export {Tabs, Tab} from './src/tabs.page';
 // export {Tooltip} from './src/tooltip.page';
